+++ src/components/ISceneLoader.h
@@ -14,18 +14,18 @@ class ISceneLoader {
     virtual void reserveShaders(size_t count) = 0;
 
     struct vec {
-        double x, y, z;
+        float x, y, z;
     };
 
     struct SceneParams {
-        vec bgColor = {0.0,0.0,0.0};
+        vec bgColor = {0.0f,0.0f,0.0f};
         std::string envMapFile = "";
     };
 
     virtual void setSceneParams(const SceneParams &sparams) = 0;
 
     struct ShaderProperty {
-        vec data = {0.0, 0.0, 0.0}; // Default to black
+        vec data = {0.0f, 0.0f, 0.0f}; // Default to black
         std::string textureName = "";
         bool hasTexture = false;
     };
@@ -37,21 +37,21 @@ class ISceneLoader {
         ShaderProperty specular;
         ShaderProperty normal;
         ShaderProperty emission;
-        double phongExp = 1.0;
-        double mirrorCoef = 0.0;
-        double roughness = 0.0;
-        vec attenuationCoef = {0.0,0.0,0.0};
-        double refractiveIndex = 1.0;
+        float phongExp = 1.0f;
+        float mirrorCoef = 0.0f;
+        float roughness = 0.0f;
+        vec attenuationCoef = {0.0f,0.0f,0.0f};
+        float refractiveIndex = 1.0f;
     };
 
     virtual void addCamera(const std::string &name, const std::string &type,
-                         const vec &pos, const vec &viewDir, double focalLength,
-                         double imagePlaneWidth) = 0;
+                         const vec &pos, const vec &viewDir, float focalLength,
+                         float imagePlaneWidth) = 0;
     virtual void addPointLight(const vec &pos, const vec &intensity) = 0;
     virtual void addAreaLight(const ISceneLoader::vec &pos,
                         const ISceneLoader::vec &color,
-                        const ISceneLoader::vec &normal, double width,
-                        double length) = 0;
+                        const ISceneLoader::vec &normal, float width,
+                        float length) = 0;
     virtual void addShader(const ISceneLoader::ShaderDesc &shaderDesc) = 0;
 
     struct ShapeDesc {
@@ -64,7 +64,7 @@ class ISceneLoader {
         std::string shaderNameReference;
 
         vec center;   // for sphere
-        double radius; // for sphere
+        float radius; // for sphere
 
         vec v0, v1, v2; // triangle vertices
