+++ src/components/light.h
@@ -5,15 +5,15 @@
 
 class Light {
     public:
-        Light(const vec3& pos, const vec3& col, double intens = 1.0)
+        Light(const vec3& pos, const vec3& col, float intens = 1.0f)
         : position(pos), color(col), intensity(intens) {}
 
         vec3 getPosition() const { return position; }
         vec3 getColor() const { return color; }
-        double getIntensity() const { return intensity; }
+        float getIntensity() const { return intensity; }
         std::string type;
     private:
         vec3 position;
         vec3 color;
-        double intensity = 1.0;
+        float intensity = 1.0f;
 };
