+++ src/components/ray.h
@@ -4,14 +4,14 @@
 
 class ray {
     public:
-        ray() : orig(0.0,0.0,0.0), dir(0.0,0.0,-1.0) {}
+        ray() : orig(0.0f,0.0f,0.0f), dir(0.0f,0.0f,-1.0f) {}
         ray(const point3& origin, const vec3& direction) : orig(origin), dir(direction) {}
         ~ray() {}
 
         const point3& origin() const {return orig;}
         const vec3& direction() const {return dir;}
 
-        point3 at(double t) const {
+        point3 at(float t) const {
             return orig+t*dir;
         }
